#18 マルチモジュール化 (#45)
* Topicの依存を解消

* :core:modelを作成

* Unitテストを削除

* Userを移動してRoomを削除

* gradle update

* data層に移動

* dataのflavorを整理

* 未追跡だった

* typo

* gradleを整理

* lint
