--- app/src/main/java/com/kabos/topicker/model/dao/UserDao.kt
@@ -1,20 +1,0 @@
-package com.kabos.topicker.model.dao
-
-import androidx.room.Dao
-import androidx.room.Insert
-import androidx.room.Query
-import androidx.room.Update
-import com.kabos.topicker.model.entity.User
-
-@Dao
-interface UserDao {
-
-    @Insert
-    suspend fun createUser(user: User)
-
-    @Query("SELECT * FROM user LIMIT 1")
-    suspend fun getUser(): User?
-
-    @Update
-    suspend fun updateFavorite(user: User)
-}
