+++ app/src/main/java/com/kabos/topicker/model/domain/TopicUiState.kt
@@ -1,7 +1,8 @@
 package com.kabos.topicker.model.domain
 
 import androidx.compose.ui.graphics.Color
-import com.kabos.topicker.model.data.Topic
+import com.kabos.model.ConversationState
+import com.kabos.model.Topic
 import com.kabos.topicker.ui.theme.*
 
 // TODO pageStateでは？？？
@@ -36,10 +37,5 @@ data class TopicUiState(
            }
        }
    }
-}
-
-sealed class ConversationState {
-    object UnSelected : ConversationState()
-    object Favorite : ConversationState()
 }
