--- app/src/main/java/com/kabos/topicker/repository/UserRepository.kt
@@ -1,10 +1,0 @@
-package com.kabos.topicker.repository
-
-import com.kabos.topicker.model.entity.User
-
-interface UserRepository {
-
-    suspend fun createUser()
-
-    suspend fun getUser(): User
-}
