+++ app/src/main/java/com/kabos/topicker/ui/viewmodel/UserViewModel.kt
@@ -3,7 +3,7 @@ package com.kabos.topicker.ui.viewmodel
 import androidx.compose.runtime.Stable
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.kabos.topicker.repository.UserRepository
+import com.kabos.domain.repository.UserRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import timber.log.Timber
