+++ core/domain/build.gradle
@@ -1,0 +1,15 @@
+plugins {
+    id 'com.android.library'
+    id 'org.jetbrains.kotlin.android'
+}
+
+android {
+    namespace 'com.kabos.domain'
+    compileSdk 32
+}
+
+dependencies {
+    implementation project(":core:model")
+
+    testImplementation 'junit:junit:4.13.2'
+}
