+++ core/domain/src/main/java/com/kabos/domain/repository/UserRepository.kt
@@ -1,0 +1,10 @@
+package com.kabos.domain.repository
+
+import com.kabos.model.User
+
+interface UserRepository {
+
+    suspend fun createUser()
+
+    suspend fun getUser(): User
+}
