+++ core/model/build.gradle
@@ -1,0 +1,13 @@
+plugins {
+    id 'com.android.library'
+    id 'org.jetbrains.kotlin.android'
+}
+
+android {
+    namespace 'com.kabos.model'
+    compileSdk 32
+}
+
+dependencies {
+
+}
