+++ core/model/src/main/java/com/kabos/model/Topic.kt
@@ -1,0 +1,23 @@
+package com.kabos.model
+
+// todo colorのmapperをどっかで書きたい String -> Color
+data class Topic(
+    val id: Int,
+    val title: String,
+    val mainColor: String, // todo これいる？
+    val conversationState: ConversationState = ConversationState.UnSelected,
+) {
+
+    fun updateConversationState(state: ConversationState): Topic =
+        Topic(
+            id = id,
+            title = title,
+            mainColor = mainColor,
+            conversationState = state,
+        )
+}
+
+sealed class ConversationState {
+    object UnSelected : ConversationState()
+    object Favorite : ConversationState()
+}
