+++ settings.gradle
@@ -14,3 +14,6 @@ dependencyResolutionManagement {
 }
 rootProject.name = "Topicker"
 include ':app'
+include ':core:model'
+include ':core:data'
+include ':core:domain'
